Attempt a delimited continuation implementation
I think to implement delimited continuations in forth you need access to
the instruction pointer as it is a part of the continuation. Because
WAForth is Subroutine threaded and WebAssembly doesn't expose the
instruction pointer we're a bit SOL.

I think it won't be too hard to modify WAForth to compile new words as
indirectly threaded and we can leave the core library as is since the
threading mechanism is going to lookup words in the table and the
call_indirect. Either it will be the traditional Forth ENTER or it will
be a core word implemented directly as a wasm function.